# docs/openapi-users.yaml
openapi: 3.0.3
info:
  title: Users API
  version: 1.0.0
  description: Dokumentasi OpenAPI 3 untuk customer module, termasuk pendaftaran, autentikasi, profil, alamat, checkout, pesanan, FAQ, tiket bantuan, chat, dan review produk.
servers:
  - url: https://haji-umroh-store-api.jauhar.id
    description: Local development server
...
paths:
...
  /api/v1/users/verify-email:
    get:
      tags:
        - User Registration
      summary: 2. Verify Email User (Customer)
      description: |
        Memverifikasi alamat email customer menggunakan token yang dikirim via email.
        Endpoint ini dipanggil ketika user mengklik link verifikasi di email.

        **Alur verifikasi:**
        1. User klik link verifikasi di email (link mengarah ke endpoint ini).
        2. Backend memvalidasi token (decode, hash, cek keberadaan, dan cek expiry).
        3. Jika valid, status user berubah dari `pending` menjadi `active`.
        4. Token ditandai sebagai consumed (tidak dapat digunakan lagi).
        5. User di-redirect ke halaman frontend berdasarkan hasil verifikasi.

        **Redirect URL:**
        - Sukses: `{FRONTEND_URL}/verify-email?status=success`
        - Token kosong: `{FRONTEND_URL}/verify-email?status=failed&reason=missing_token`
        - Token invalid/expired: `{FRONTEND_URL}/verify-email?status=failed&reason=invalid_or_expired`
        - Error lainnya: `{FRONTEND_URL}/verify-email?status=failed&reason=invalid_token`

        **Catatan:** Endpoint ini tidak mengembalikan JSON response, melainkan melakukan HTTP redirect (302 Found).
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: query
          required: true
          description: Token verifikasi yang dikirim melalui email (base64url encoded).
          schema:
            type: string
          example: dGhpcyBpcyBhIHNhbXBsZSB0b2tlbiB2YWx1ZQ==
      responses:
        "302":
          description: |
            Redirect ke halaman frontend. Lokasi redirect tergantung hasil verifikasi:
            - **Sukses** → `{FRONTEND_URL}/verify-email?status=success`
            - **Gagal (token kosong)** → `{FRONTEND_URL}/verify-email?status=failed&reason=missing_token`
            - **Gagal (token invalid/expired)** → `{FRONTEND_URL}/verify-email?status=failed&reason=invalid_or_expired`
            - **Gagal (error lainnya)** → `{FRONTEND_URL}/verify-email?status=failed&reason=invalid_token`
          headers:
            Location:
              description: URL tujuan redirect di frontend.
              schema:
                type: string
                example: https://app.example.com/verify-email?status=success
...
  /api/v1/users/resend-verification:
    post:
      tags:
        - User Registration
      summary: Resend Email Verifikasi
      description: |
        Mengirim ulang email verifikasi ke alamat email yang didaftarkan.
        Endpoint ini dirancang untuk keamanan — response selalu sukses meskipun email tidak ditemukan
        (untuk mencegah enumerasi email).

        **Aturan:**
        - User harus berstatus `pending` (belum diverifikasi).
        - Terdapat cooldown **1 menit** antara setiap permintaan kirim ulang.
        - Token verifikasi lama akan di-invalidate sebelum token baru dibuat.
        - Link verifikasi baru berlaku selama **24 jam**.

        **Perilaku khusus:**
        - Jika email tidak ditemukan → response tetap sukses (keamanan).
        - Jika user sudah aktif → error 409 (Conflict).
        - Jika user bukan status `pending` → error 400 (Bad Request).
        - Jika request terlalu cepat (< 1 menit) → error 429 (Too Many Requests).
      operationId: resendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationRequest"
            example:
              email: ahmad.fauzi@example.com
      responses:
        "200":
          description: Permintaan kirim ulang diterima. Jika email terdaftar dan masih pending, email verifikasi baru telah dikirim.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResendVerificationEnvelope"
              example:
                success: true
                message: "If your email is registered and pending verification, a new verification email has been sent."
                data: null
                errors: null
                meta: null
        "400":
          description: Request tidak valid (validasi gagal atau user bukan status pending).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                missingEmail:
                  summary: Email wajib diisi
                  value:
                    success: false
                    message: "Invalid request body"
                    data: null
                    errors: "Key: 'ResendVerificationRequest.Email' Error:Field validation for 'Email' failed on the 'required' tag"
                    meta: null
                invalidEmailFormat:
                  summary: Format email tidak valid
                  value:
                    success: false
                    message: "Invalid request body"
                    data: null
                    errors: "Key: 'ResendVerificationRequest.Email' Error:Field validation for 'Email' failed on the 'email' tag"
                    meta: null
                userNotPending:
                  summary: User bukan status pending
                  value:
                    success: false
                    message: "user is not in pending status"
                    data: null
                    errors: null
                    meta: null
        "409":
          description: Konflik — user sudah aktif (sudah diverifikasi).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "user is already active"
                data: null
                errors: null
                meta: null
        "429":
          description: Terlalu banyak permintaan — silakan tunggu sebelum meminta email verifikasi lagi.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "please wait before requesting another verification email"
                data: null
                errors: null
                meta: null
        "500":
          description: Internal server error.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "An unexpected error occurred"
                data: null
                errors: null
                meta: null
...
  /api/v1/users/reset-password:
    post:
      tags:
        - Password Recovery
      summary: 3. Reset Password (Forgot Password)
      description: |
        Mengatur ulang password user setelah OTP diverifikasi.
        Endpoint ini menerima `proof_token` yang didapat dari verifikasi OTP
        dan password baru yang memenuhi syarat kekuatan.

        **Alur lengkap forgot password:**
        1. `POST /api/v1/otp/request` — kirim OTP ke email.
        2. `POST /api/v1/otp/verify` — verifikasi kode OTP → dapat `proof_token`.
        3. `POST /api/v1/users/reset-password` — reset password dengan `proof_token`.

        **Syarat password baru:**
        - Minimal 8 karakter.
        - Mengandung minimal 1 huruf besar.
        - Mengandung minimal 1 huruf kecil.
        - Mengandung minimal 1 angka.
        - Mengandung minimal 1 karakter spesial.

        **Efek samping:**
        - Semua sesi aktif (JWT token lama) akan otomatis di-invalidasi setelah password berhasil direset.
        - User perlu login ulang setelah reset password.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            example:
              proof_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              new_password: NewSecurePass456@
      responses:
        "200":
          description: Password berhasil direset. User perlu login ulang.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessNullDataEnvelope"
              example:
                success: true
                message: "password reset successfully"
                data: null
                errors: null
                meta: null
        "400":
          description: Request tidak valid (validasi gagal, proof token invalid, atau password terlalu lemah).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidProofToken:
                  summary: Proof token tidak valid atau expired
                  value:
                    success: false
                    message: "invalid or expired proof token"
                    data: null
                    errors: null
                    meta: null
                weakPassword:
                  summary: Password terlalu lemah
                  value:
                    success: false
                    message: "password does not meet strength requirements: password must contain at least one uppercase letter"
                    data: null
                    errors: null
                    meta: null
                missingFields:
                  summary: Field wajib tidak diisi
                  value:
                    success: false
                    message: "Invalid request body"
                    data: null
                    errors: "Key: 'ResetPasswordRequest.ProofToken' Error:Field validation for 'ProofToken' failed on the 'required' tag"
                    meta: null
        "404":
          description: User tidak ditemukan.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "user not found"
                data: null
                errors: null
                meta: null
        "500":
          description: Internal server error.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "An unexpected error occurred"
                data: null
                errors: null
                meta: null
  /api/v1/users/change-password:
    put:
      tags:
        - User Profile
      summary: Change Password
      description: |
        Mengubah password user yang sedang login.
        User harus memasukkan password lama yang benar dan password baru yang memenuhi syarat kekuatan.

        **Endpoint ini tersedia untuk semua role:** `customer`, `umkm`, `admin`, `cs`, `finance`.

        **Syarat:**
        - Wajib login (menyertakan JWT token).
        - Password lama harus benar.
        - Password baru harus berbeda dari password lama.
        - Password baru minimal 8 karakter, mengandung huruf besar, huruf kecil, angka, dan karakter spesial.

        **Efek samping:**
        - Semua sesi aktif (JWT token lama) akan otomatis di-invalidasi setelah password berhasil diubah.
        - User perlu login ulang setelah ganti password.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
            example:
              old_password: OldSecurePass123!
              new_password: NewSecurePass456@
      responses:
        "200":
          description: Password berhasil diubah. User perlu login ulang.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessNullDataEnvelope"
              example:
                success: true
                message: "password changed successfully"
                data: null
                errors: null
                meta: null
        "400":
          description: Request tidak valid (password lama salah, password baru sama, atau password terlalu lemah).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                oldPasswordIncorrect:
                  summary: Password lama salah
                  value:
                    success: false
                    message: "old password is incorrect"
                    data: null
                    errors: null
                    meta: null
                samePassword:
                  summary: Password baru sama dengan password lama
                  value:
                    success: false
                    message: "new password must be different from old password"
                    data: null
                    errors: null
                    meta: null
                weakPassword:
                  summary: Password terlalu lemah
                  value:
                    success: false
                    message: "password does not meet strength requirements: password must contain at least one special character"
                    data: null
                    errors: null
                    meta: null
                missingFields:
                  summary: Field wajib tidak diisi
                  value:
                    success: false
                    message: "Invalid request body"
                    data: null
                    errors: "Key: 'ChangePasswordRequest.OldPassword' Error:Field validation for 'OldPassword' failed on the 'required' tag"
                    meta: null
        "401":
          description: Token tidak diberikan, tidak valid, atau sudah expired.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                missingHeader:
                  summary: Header Authorization tidak ada
                  value:
                    success: false
                    message: "missing authorization header"
                    data: null
                    errors: null
                    meta: null
                invalidToken:
                  summary: Token tidak valid atau expired
                  value:
                    success: false
                    message: "invalid or expired token"
                    data: null
                    errors: null
                    meta: null
                sessionExpired:
                  summary: Sesi expired karena password diubah
                  value:
                    success: false
                    message: "session expired due to password change, please login again"
                    data: null
                    errors: null
                    meta: null
        "403":
          description: User tidak memiliki role yang diizinkan.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "insufficient permissions"
                data: null
                errors: null
                meta: null
        "404":
          description: User tidak ditemukan.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "user not found"
                data: null
                errors: null
                meta: null
        "500":
          description: Internal server error.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "An unexpected error occurred"
                data: null
                errors: null
                meta: null
  /api/v1/users/delete-account:
    post:
      tags:
        - User Profile
      summary: Delete Account (Deactivate)
      description: |
        Menonaktifkan akun customer yang sedang login (soft delete).
        Akun akan diubah statusnya menjadi `deactivated` dan tidak dapat login lagi.
        Data akun **tidak dihapus** dari database, hanya dinonaktifkan.

        **Endpoint ini hanya tersedia untuk role:** `customer`.

        **Syarat:**
        - Wajib login (menyertakan JWT token).
        - Wajib memilih alasan penghapusan.
        - Wajib mengonfirmasi dengan memasukkan password yang benar.

        **Efek:**
        - Status akun berubah menjadi `deactivated`.
        - Alasan dan waktu permintaan penghapusan disimpan.
        - User tidak dapat login lagi setelah akun dinonaktifkan.
        - Tindakan ini tidak dapat dibatalkan oleh user sendiri.
      operationId: deleteAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
            example:
              reason: "Tidak menggunakan lagi"
              password: "SecurePass123!"
      responses:
        "200":
          description: Akun berhasil dinonaktifkan.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessNullDataEnvelope"
              example:
                success: true
                message: "account has been deactivated"
                data: null
                errors: null
                meta: null
        "400":
          description: Request tidak valid (password salah atau field wajib tidak diisi).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                passwordIncorrect:
                  summary: Password salah
                  value:
                    success: false
                    message: "password is incorrect"
                    data: null
                    errors: null
                    meta: null
                missingFields:
                  summary: Field wajib tidak diisi
                  value:
                    success: false
                    message: "Invalid request body"
                    data: null
                    errors: "Key: 'DeleteAccountRequest.Reason' Error:Field validation for 'Reason' failed on the 'required' tag"
                    meta: null
        "401":
          description: Token tidak diberikan, tidak valid, atau sudah expired.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                missingHeader:
                  summary: Header Authorization tidak ada
                  value:
                    success: false
                    message: "missing authorization header"
                    data: null
                    errors: null
                    meta: null
                invalidToken:
                  summary: Token tidak valid atau expired
                  value:
                    success: false
                    message: "invalid or expired token"
                    data: null
                    errors: null
                    meta: null
        "409":
          description: Akun sudah dinonaktifkan sebelumnya.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "account is already deactivated"
                data: null
                errors: null
                meta: null
        "500":
          description: Internal server error.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIDHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: "internal server error"
                data: null
                errors: null
                meta: null
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token yang didapat dari endpoint login.
  headers:
    RequestIDHeader:
      description: Unique request ID untuk tracing dan debugging.
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
...
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Alamat email yang terdaftar.
          example: ahmad.fauzi@example.com
...
    ChangePasswordRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          format: password
          description: Password lama yang saat ini digunakan.
          example: OldSecurePass123!
        new_password:
          type: string
          format: password
          minLength: 8
          description: |
            Password baru. Harus memenuhi syarat kekuatan:
            - Minimal 8 karakter.
            - Mengandung minimal 1 huruf besar.
            - Mengandung minimal 1 huruf kecil.
            - Mengandung minimal 1 angka.
            - Mengandung minimal 1 karakter spesial.
          example: NewSecurePass456@
    DeleteAccountRequest:
      type: object
      required:
        - reason
        - password
      properties:
        reason:
          type: string
          maxLength: 50
          description: |
            Alasan penghapusan akun. Pilihan yang tersedia:
            - Tidak menggunakan lagi
            - Privasi dan keamanan
            - Pengalaman buruk
            - Beralih ke platform lain
            - Lainnya
          example: Tidak menggunakan lagi
        password:
          type: string
          format: password
          description: Password akun untuk konfirmasi penghapusan.
          example: SecurePass123!
    ResetPasswordRequest:
      type: object
      required:
        - proof_token
        - new_password
      properties:
        proof_token:
          type: string
          description: Proof token JWT yang didapat dari endpoint `POST /api/v1/otp/verify` setelah verifikasi OTP berhasil.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        new_password:
          type: string
          format: password
          minLength: 8
          description: |
            Password baru. Harus memenuhi syarat kekuatan:
            - Minimal 8 karakter.
            - Mengandung minimal 1 huruf besar.
            - Mengandung minimal 1 huruf kecil.
            - Mengandung minimal 1 angka.
            - Mengandung minimal 1 karakter spesial.
          example: NewSecurePass456@
...
    SuccessResendVerificationEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "If your email is registered and pending verification, a new verification email has been sent."
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: object
          nullable: true
          example: null
        meta:
          type: object
          nullable: true
          example: null
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Pesan error.
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: object
          nullable: true
          description: Detail error tambahan (biasanya string validasi error).
        meta:
          type: object
          nullable: true
          example: null

    # ──────────── Login Response Schemas ────────────
...
    SuccessNullDataEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: object
          nullable: true
          example: null
        meta:
          type: object
          nullable: true
          example: null
...